+++ src/test/java/racingcar/CarTest.java
@@ -3,6 +3,7 @@ package racingcar;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
 import racingcar.util.CarStatusSetterForTest;
 
 import static org.assertj.core.api.Assertions.assertThat;
