+++ src/test/java/racingcar/CarsTest.java
@@ -2,7 +2,7 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.util.CarStatusSetterForTest;
+import racingcar.domain.Cars;
 
 import static org.assertj.core.api.Assertions.assertThat;
